refactor: omul read actions
