+++ src/app/ombul/page.tsx
@@ -3,7 +3,7 @@ import CreateOmbul from './CreateOmbul'
 import OmbulCover from './OmbulCover'
 import PageWrapper from '@/components/PageWrapper/PageWrapper'
 import AddHeaderItemPopUp from '@/components/AddHeaderItem/AddHeaderItemPopUp'
-import { readLatestOmbul, readOmbuls } from '@/actions/ombul/read'
+import { readLatestOmbulAction, readOmbulsAction } from '@/actions/ombul/read'
 import { getUser } from '@/auth/user'
 import { ExpandedOmbul } from '@/server/ombul/Types'
 
@@ -15,9 +15,9 @@ export default async function Ombuls() {
 
     const showCreateButton = user.permissions.includes('OMBUL_CREATE')
 
-    const latestOmbulRes = await readLatestOmbul()
+    const latestOmbulRes = await readLatestOmbulAction()
     const latestOmbul = latestOmbulRes.success ? latestOmbulRes.data : null
-    const ombulRes = await readOmbuls()
+    const ombulRes = await readOmbulsAction()
     if (!ombulRes.success) throw new Error('Failed to read ombuls')
     const ombuls = ombulRes.data
